+++ CHANGELOG.md
@@ -3,7 +3,9 @@ Version 0.21
 - shortened json keys (url,token,text,old,type) -> (a,b,c,o,t)
 - include read_access flag in datalink (r)
 - include mapping flag in datalink (m)
-- new view datalink.access with triggers to update datalink.directory.diracl
+- new view `datalink.access` with triggers to update `datalink.directory.diracl`
+- new function `datalink.has_file_privilege()` to enquire about file privileges
+- initial `datalink.insight` table for read access management
 
 Version 0.20
 ------------